interface was using floats still on accident, fixed shapes not iterating position

+++ extensions/multiplacement/src/_aux.c
@@ -97,5 +97,5 @@ double shape_average(double array[], int num_elements){
   for (int i = 0; i < num_elements; i++){
     sum += array[i];
   }
-  return sum/(num_elements + 2);
+  return sum/((double)(num_elements + 2));
 }
